pyproject.toml
[tool.poetry]
name = "dj-cors-example"
version = "0.1.0"
description = ""
authors = ["shkumagai <take.this.2.your.grave@gmail.com>"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
isort = "^5.7.0"
black = "^20.8b1"
mypy = "^0.812"
tox = "^3.23.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
